+++ src/main/java/com/team5/projrental/product/model/review/ReviewGetDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReviewGetDto {
-    private Integer iproduct;
+    private Long iproduct;
     private Integer page;
     private Integer prodPerPage;
 }
